// types/certi.ts
export type certificatDbType = {
  collectiv_parent: string;
  date_deliv_cert: string;
  date_nais_enfant: string;
  date_nais_mere: string;
  date_nais_pere: string;
  hospital_id: number;
  id: number;
  lieu_nais_mere: string;
  lieu_nais_pere: string;

  localite_parent: string;

  nationalite_mere: string;

  nationalite_pere: string;

  nom_complet_mere: string;

  nom_complet_pere: string;
  nom_enfant: string;
  nom_medecin: string;
  poid_enfant: string;
  post_nom_enfant: string;
  prenom_enfant: string;
  profession_mere: string;
  profession_pere: string;
  sexe_enfant: string;
};